+++ hbs2-hub/app/Main.hs
@@ -17,6 +17,7 @@ import HBS2.Hub.CLI.Maintainer
 import HBS2.Hub.CLI.Own
 import HBS2.Hub.CLI.Policy
 import HBS2.Hub.CLI.Pr
+import HBS2.Hub.CLI.Publish
 import HBS2.Hub.CLI.Read
 import HBS2.Hub.CLI.Show
 import HBS2.Hub.CLI.Reject
@@ -124,6 +125,7 @@ main = do
         readEntries
         showEntries
         prEntries
+        publishEntries
         maintainerEntries
         rejectEntries
         ownEntries
@@ -280,7 +282,7 @@ main = do
                     , "hub:log", "hub:maintainer:list", "hub:ban:list"
                     , "hub:ban", "hub:unban"
                     , "hub:maintainer:add", "hub:maintainer:remove"
-                    , "hub:pr:merge", "hub:pr:checkout", "hub:sync"
+                    , "hub:pr:merge", "hub:pr:checkout", "hub:sync", "hub:publish"
                     , "hub:issue:close", "hub:issue:reopen", "hub:issue:label"
                     , "hub:issue:assign", "hub:compact"
                     , "hub:redact"